requirements.txt
# YouTube Video Creator Bot - Dependencies

# AI Providers
openai>=1.0.0           # OpenAI (ChatGPT) and Grok (xAI uses OpenAI-compatible API)
google-generativeai>=0.3.0  # Google Gemini

# Browser automation
playwright>=1.40.0

# HTTP client for downloads
httpx>=0.25.0

# Type hints (for Python < 3.10)
typing-extensions>=4.0.0
